share spell effects
