+++ packages/shared/src/entities/index.ts
@@ -1,4 +1,5 @@
 export * from './battle-state';
+export * from './spell-effect';
 export * from './Character';
 export * from './Player';
 export * from './Spell';
